1. removed gallery wrapper and simplified galleryhome to do only intended action in userstory 2. suspense wrapper as part of galleryhome component 3. removed skeleton container and made it as reusable mixin 4. removed redundant errorboundary wrapper component 5. simplified image rendering logic to only fetch and render full sized images. no resize handlers.
