--- src/components/Gallery/components/GalleryPicture/GalleryPicture.constants.ts
@@ -1,1 +1,0 @@
-export const PICTURE_MAX_WIDTH = 270; // px
