--- src/components/Gallery/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useGallery';
