+++ src/components/GalleryPicture/index.ts
