+++ src/components/GallerySkeleton/index.ts
