+++ src/components/ui/ErrorFallbackButton/ErrorFallbackButton.constants.ts
@@ -1,0 +1,3 @@
+export const ERROR_FALLBACK_BUTTON_HEADING = 'Something Went Wrong';
+export const ERROR_FALLBACK_BUTTON_DESCRIPTION = 'Click to try again';
+export const ERROR_FALLBACK_BUTTON_TITLE = 'Error! Try again';
