+++ src/components/ui/ErrorFallbackButton/ErrorFallbackButton.tsx
@@ -2,6 +2,11 @@ import ErrorLogo from '@assets/sad-cloud.svg?react';
 import classNames from 'classnames/bind';
 import type { ReactElement } from 'react';
 import styles from './ErrorFallbackButton.module.scss';
+import {
+  ERROR_FALLBACK_BUTTON_DESCRIPTION,
+  ERROR_FALLBACK_BUTTON_HEADING,
+  ERROR_FALLBACK_BUTTON_TITLE,
+} from './ErrorFallbackButton.constants';
 
 const cx = classNames.bind(styles);
 
@@ -14,9 +19,9 @@ type ErrorFallbackButtonProps = {
 
 const ErrorFallbackButton = ({
   onClick,
-  heading = 'Something Went Wrong',
-  description = 'Click to try again',
-  title = 'Error! Try again',
+  heading = ERROR_FALLBACK_BUTTON_HEADING,
+  description = ERROR_FALLBACK_BUTTON_DESCRIPTION,
+  title = ERROR_FALLBACK_BUTTON_TITLE,
 }: ErrorFallbackButtonProps): ReactElement => (
   <button type="button" onClick={onClick} title={title} aria-label={title} className={cx('error-fallback-button')}>
     <ErrorLogo className={cx('error-fallback-button__logo')} />
