+++ src/hooks/useGallery.types.ts
