+++ src/pages/HomePage/HomePage.tsx
@@ -1,10 +1,6 @@
-import Gallery from '@components/Gallery';
+import GalleryHome from '@components/GalleryHome';
 import type { ReactElement } from 'react';
 
-const HomePage = (): ReactElement => (
-  <Gallery>
-    <Gallery.Home />
-  </Gallery>
-);
+const HomePage = (): ReactElement => <GalleryHome />;
 
 export default HomePage;
