+++ src/scss/mixins/mixins.scss
@@ -1,0 +1,16 @@
+@use '/src/scss/variables/variables.scss' as *;
+
+@mixin skeleton-animation($start: var(--skeleton-start-bg), $end: var(--skeleton-end-bg)) {
+  animation: skeleton-loading 1s linear infinite alternate;
+  color: transparent;
+
+  @keyframes skeleton-loading {
+    0% {
+      background-color: #{$start};
+    }
+
+    100% {
+      background-color: #{$end};
+    }
+  }
+}
